Update Fitur Peminjaman

+++ app/Http/Controllers/AlatBahanController.php
@@ -2,6 +2,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\AlatBahan;
+use App\Models\Peminjaman;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 
@@ -12,7 +13,7 @@ class AlatBahanController extends Controller
         $alat = AlatBahan::paginate(10);
         return Inertia::render('AlatBahan/Index', [
             'alat' => $alat,
-            'flash' => session()->get('flash') // Pastikan ini ada
+            'flash' => session()->get('flash')
         ]);
     }
 
@@ -28,7 +29,7 @@ class AlatBahanController extends Controller
             'jenis' => 'required|in:alat,bahan',
             'kondisi' => 'required|in:baik,rusak,baru,hilang',
             'jumlah' => 'required|integer|min:0',
-            'deskripsi' => 'required|string',
+            'deskripsi' => 'nullable|string',
         ]);
 
         AlatBahan::create($validated);
@@ -51,7 +52,7 @@ class AlatBahanController extends Controller
             'jenis' => 'required|string',
             'kondisi' => 'required|in:baik,rusak,hilang,baru',
             'jumlah' => 'required|integer|min:1',
-            'deskripsi' => 'string',
+            'deskripsi' => 'nullable|string',
         ]);
 
         $alat->update($request->all());
@@ -64,7 +65,23 @@ class AlatBahanController extends Controller
 
     public function destroy(AlatBahan $alat)
     {
+        $isBorrowed = Peminjaman::where('id_alat', $alat->id_alat)
+            ->where('status', 'dipinjam')
+            ->exists();
+
+        if ($isBorrowed) {
+            return redirect()->route('alat.index')
+                ->with('error', 'Tidak dapat menghapus. Alat/Bahan sedang dipinjam!');
+        }
+
+        if ($alat->jenis === 'bahan') {
+            return redirect()->route('alat.index')
+                ->with('error', 'Bahan tidak dapat dihapus/dikembalikan!');
+        }
+
         $alat->delete();
-        return redirect()->route('alat.index')->with('success', 'Alat berhasil dihapus.');
+        
+        return redirect()->route('alat.index')
+            ->with('success', 'Alat berhasil dihapus.');
     }
 }
